+++ kaldigstserver/server/Application.py
@@ -1,0 +1,47 @@
+import tornado.web
+import json
+import os
+
+
+class Application(tornado.web.Application):
+    def __init__(self):
+        settings = dict(
+            cookie_secret="43oETzKXQAGaYdkL5gEmGeJJFuYh7EQnp2XdTP1o/Vo=",
+            template_path=os.path.join(os.path.dirname(os.path.dirname(__file__)), "templates"),
+            static_path=os.path.join(os.path.dirname(os.path.dirname(__file__)), "static"),
+            xsrf_cookies=False,
+            autoescape=None,
+        )
+
+        handlers = [
+            (r"/", MainHandler),
+            (r"/client/ws/speech", DecoderSocketHandler),
+            (r"/client/ws/status", StatusSocketHandler),
+            (r"/client/dynamic/reference", ReferenceHandler),
+            (r"/client/dynamic/recognize", HttpChunkedRecognizeHandler),
+            (r"/worker/ws/speech", WorkerSocketHandler),
+            (r"/client/static/(.*)", tornado.web.StaticFileHandler, {'path': settings["static_path"]}),
+        ]
+        tornado.web.Application.__init__(self, handlers, **settings)
+        self.available_workers = set()
+        self.status_listeners = set()
+        self.num_requests_processed = 0
+
+    def send_status_update_single(self, ws):
+        status = dict(num_workers_available=len(self.available_workers), num_requests_processed=self.num_requests_processed)
+        ws.write_message(json.dumps(status))
+
+    def send_status_update(self):
+        for ws in self.status_listeners:
+            self.send_status_update_single(ws)
+
+    def save_reference(self, content_id, content):
+        refs = {}
+        try:
+            with open("reference-content.json") as f:
+                refs = json.load(f)
+        except:
+            pass
+        refs[content_id] = content
+        with open("reference-content.json", "w") as f:
+            json.dump(refs, f, indent=2)
