+++ kaldigstserver/server/StatusSocketHandler.py
@@ -1,0 +1,17 @@
+import tornado.websocket
+import logging
+
+
+class StatusSocketHandler(tornado.websocket.WebSocketHandler):
+    # needed for Tornado 4.0
+    def check_origin(self, origin):
+        return True
+
+    def open(self):
+        logging.info("New status listener")
+        self.application.status_listeners.add(self)
+        self.application.send_status_update_single(self)
+
+    def on_close(self):
+        logging.info("Status listener left")
+        self.application.status_listeners.remove(self)
